requirements.txt
# Core AI/ML Libraries
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
textblob>=0.17.1
vaderSentiment>=3.3.2

# Advanced AI Models (Optional but recommended)
transformers>=4.15.0
torch>=1.10.0

# Web Framework
flask>=2.0.0
flask-cors>=3.0.10

# Database
pymongo>=4.0.0

# Data Processing
requests>=2.28.0
beautifulsoup4>=4.11.0
geopy>=2.2.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
wordcloud>=1.8.0

# Utilities
python-dotenv>=0.19.0
schedule>=1.1.0

# Social Media APIs (Optional)
tweepy>=4.12.0
snscrape>=0.4.3

# Development and Testing
pytest>=6.0.0
pytest-cov>=3.0.0

# Production
gunicorn>=20.1.0


pymongo